requirements.txt
# Core
requests
pause

# Waveshare
pillow
spidev
RPi.GPIO